# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.CreateUserPayload:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
    required:
    - name
    - password
    - role
    - status
    type: object
...
  admin.UpdateUserPayload:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
    required:
    - name
    - role
    - status
    type: object
...
info:
  contact: {}
  description: GameLink 平台 API，包含健康检查、认证与管理端能力。
  title: GameLink API
  version: 0.3.0
paths:
...
  /admin/users:
    get:
      description: 分页列出用户，支持按角色/状态/时间范围/关键字筛选
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - collectionFormat: csv
        description: 角色过滤，可多值
        in: query
        items:
          type: string
        name: role
        type: array
      - collectionFormat: csv
        description: 状态过滤，可多值
        in: query
        items:
          type: string
        name: status
        type: array
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      - description: 关键字（匹配 name/email/phone）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出用户
      tags:
      - Admin/Users
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - Admin/Users
  /admin/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - Admin/Users
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户
      tags:
      - Admin/Users
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - Admin/Users
  /admin/users/{id}/logs:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 动作过滤
        enum:
        - create
        - update
        - delete
        in: query
        name: action
        type: string
      - description: 操作者用户ID
        in: query
        name: actor_user_id
        type: integer
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      - description: 导出格式
        enum:
        - csv
        in: query
        name: export
        type: string
      - description: 导出列（逗号分隔）
        in: query
        name: fields
        type: string
      - description: 列头语言
        enum:
        - en
        - zh
        in: query
        name: header_lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户操作日志
      tags:
      - Admin/Users
  /admin/users/{id}/orders:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - collectionFormat: csv
        description: 订单状态
        in: query
        items:
          type: string
        name: status
        type: array
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户的订单
      tags:
      - Admin/Users
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{role}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户角色
      tags:
      - Admin/Users
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{status}'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - Admin/Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"